backend/telegrambot/requirements.txt
python-telegram-bot>=21
httpx>=0.27
python-dotenv>=1.0
rapidfuzz>=3.9
fastapi==0.115.*
uvicorn==0.30.*
gunicorn==21.*